% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalcrime_function.R
\name{analyze_temporal_trends}
\alias{analyze_temporal_trends}
\title{Analyze Temporal Trends in Crime Incidents}
\usage{
analyze_temporal_trends(crime_data)
}
\arguments{
\item{crime_data}{A data frame containing the crime dataset.}
}
\value{
A list containing visualizations for the yearly trends, along with summary statistics for years and months.
}
\description{
This function analyzes the temporal trends in crime incidents based on year and month. It generates plots
to visualize the yearly trends and monthly trends for specific years.
}
\examples{
# Load the Vancouver crime dataset
crime_data <- load_vancouver_crime_data()

# Analyze temporal trends
temporal_trends <- analyze_temporal_trends(crime_data)
print(temporal_trends$text_results)
#print(temporal_trends$year_plot)
for (result in temporal_trends$specific_year_month_results) {
  cat(result$month_text_results, "\n")
  #print(result$month_plot)
}

}
