% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood_visualization.R
\name{calculate_total_crime_count_by_type_neighborhood}
\alias{calculate_total_crime_count_by_type_neighborhood}
\title{Calculate Total Crime Count by Type in a Neighborhood for a Specific Year and Month}
\usage{
calculate_total_crime_count_by_type_neighborhood(
  crime_data,
  neighborhood,
  year,
  month
)
}
\arguments{
\item{crime_data}{A data frame containing the crime dataset.}

\item{neighborhood}{A character string representing the neighborhood.}

\item{year}{A numeric value representing the year.}

\item{month}{A numeric value representing the month.}
}
\value{
A data frame with columns: TYPE and count.
}
\description{
Calculate the total number of crimes for each crime type in a specific neighborhood
for a given year and month.
}
\examples{
# Load the crime dataset
crime_data <- load_vancouver_crime_data()

# Calculate total crime count by type in a neighborhood for a specific year and month
crime_count_by_type_neighborhood <- calculate_total_crime_count_by_type_neighborhood(crime_data,
                                                                                       neighborhood = "Central Business District",
                                                                                       year = 2017,
}
