% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theft_type_temporal.R
\name{find_highest_theft_type}
\alias{find_highest_theft_type}
\title{Find the Theft Type with the Highest Count}
\usage{
find_highest_theft_type(theft_data, years, max_months)
}
\arguments{
\item{theft_data}{A data frame containing the theft dataset.}

\item{years}{A numeric vector of years.}

\item{max_months}{A character vector of maximum crime count months (abbreviated).}
}
\value{
A list containing the highest theft type for each year and month.
}
\description{
For the specified years and months, this function finds the theft type with the
highest count in the provided theft dataset.
}
\examples{
# Load the theft dataset
theft_data <- load_vancouver_crime_data()

# Find highest theft type for specific years and months
highest_theft_types <- find_highest_theft_type(theft_data, c(2003, 2015, 2017), c("May", "Nov", "May"))
# Print the highest theft types for each year and month
for (key in names(highest_theft_types)) {
year_month <- key
highest_type <- highest_theft_types[[key]]$TYPE
count <- highest_theft_types[[key]]$count

cat("Year and Month:", year_month, "\n")
cat("Highest Theft Type:", highest_type, "\n")
cat("Count:", count, "\n\n")
}
}
