% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_theft_type.R
\name{find_least_theft_type}
\alias{find_least_theft_type}
\title{Find the Theft Type with the Least Count}
\usage{
find_least_theft_type(theft_data, years, max_months)
}
\arguments{
\item{theft_data}{A data frame containing the theft dataset.}

\item{years}{A numeric vector of years.}

\item{max_months}{A character vector of maximum crime count months (abbreviated).}
}
\value{
A list containing the least theft type for each year and month.
}
\description{
For the specified years and months, this function finds the theft type with the
least count in the provided theft dataset.
}
\examples{
# Load the theft dataset
theft_data <- load_vancouver_crime_data()

# Find least theft type for specific years and months
least_theft_types <- find_least_theft_type(theft_data, c(2003, 2015, 2017), c("May", "Nov", "May"))
# Print the least theft types for each year and month
for (key in names(least_theft_types)) {
 year_month <- key
 least_type <- least_theft_types[[key]]$TYPE
 count <- least_theft_types[[key]]$count

 cat("Year and Month:", year_month, "\n")
 cat("Least Theft Type:", least_type, "\n")
 cat("Count:", count, "\n\n")
}
}
